+++ routes/post.js
@@ -1,7 +1,6 @@
 // 게시글 관련 라우터
 const express = require("express");
 const controller = require("../controller/Cpost");
-// const uploadDetail = require("../middlewares/uploadDetail");
 const router = express.Router();
 
 // GET /posts/list
@@ -20,7 +19,6 @@ router.get("/search/:category", controller.getPostsByKeywordByCategory);
 router.get("/", controller.getPosts);
 
 // POST /posts
-// router.post("/", uploadDetail.single("file"), controller.postPosts);
 router.post("/", controller.postPosts);
 
 // GET /posts/detail/:p_seq
